requirements.txt
alembic==1.12.1
amqp==5.2.0
billiard==4.2.0
blinker==1.7.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Faker==20.1.0
Flask==3.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.4
Mako==1.3.0
MarkupSafe==2.1.3
prompt-toolkit==3.0.41
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
tzdata==2023.3
vine==5.1.0
wcwidth==0.2.12
Werkzeug==3.0.1
